Refactor Registration Service and Controller into CRUD Services
