+++ src/main/java/de/fsr/mariokart_backend/registration/controller/RegistrationUpdateController.java
@@ -1,0 +1,37 @@
+package de.fsr.mariokart_backend.registration.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import de.fsr.mariokart_backend.exception.EntityNotFoundException;
+import de.fsr.mariokart_backend.registration.model.dto.TeamInputDTO;
+import de.fsr.mariokart_backend.registration.model.dto.TeamReturnDTO;
+import de.fsr.mariokart_backend.registration.service.RegistrationUpdateService;
+import lombok.AllArgsConstructor;
+
+@CrossOrigin
+@RestController
+@RequestMapping("/teams")
+@AllArgsConstructor
+public class RegistrationUpdateController {
+
+    private final RegistrationUpdateService registrationUpdateService;
+
+    @PutMapping("/{id}")
+    @ResponseBody
+    public ResponseEntity<TeamReturnDTO> updateTeam(@PathVariable Long id, @RequestBody TeamInputDTO teamCreation) {
+        try {
+            return ResponseEntity.ok(registrationUpdateService.updateTeam(id, teamCreation));
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().build();
+        }
+    }
+}
